in progress work...
working on the forms example, but it's a mess.

The concept of splitting the action from the view worked well enough
without form validation, but with it, I'm seeing several variations of
load that have to happen at different times (some before the event some after the event)

Right now, this is a mess... I have to think long and hard as to if it can
be cleaned up.  If not, then perhaps I'm forced into doing things manually in Rails
and abandoning the state machine idea.  It won't go down without a fight though!

The vision is to be able to simply change the state machine and have user flows
change behavior (i.e. "just works") vs. the Rails way, where
you often have to refactor the entire user flow every time you make a change unless you
are extremely consistent about how you code it.  I'm thinking that maybe I've
uncovered the reason why doing this manually the Rails way is so difficult.

can it be contained? needs more analysis.