* converted State into Hashie::Mash * guard becomes def_delegator to state
I found in practice that I was writing a lot of duplicate code in the before_view and the guard and the state.
This simplifies the code by letting one definition (in the form) govern auto sync to state,
plus any guards can be added on the fly to state and accessed directly from the aasm block by using
Hashie::Mash's auto '?' methods -- i.e. state.something = true automatically gets state.something? == true
This minimizes the number of duplicate properties floating around as I try to manage state between the controller
and the form objects and the guards.