updating form example showing how to marshal State to hash on load and save. this lets the session store a plain Ruby Hash which has maximum portability, yet lets the controller leverage the Mash helpers and ease of state management.
I tried to find a way to do this automatically under the covers, but it seems
complicated.  For now, I'll demonstrate it.  It's separate from the actual
StatefulController idea anyway, since if you were using other stores, you'd
have to do custom marshalling to and from your State object anyway, so maybe that's right.